+++ concierge-elastic/src/main/java/com/grookage/concierge/elastic/repository/ElasticRepository.java
@@ -51,6 +51,7 @@ public class ElasticRepository implements ConciergeRepository {
     private static final String CONFIG_NAME = "configName";
     private static final String VERSION = "version";
     private static final String CONFIG_STATE = "configState";
+    private static final String CONFIG_TYPE = "configType";
     private final ElasticsearchClient client;
     private final ElasticConfig elasticConfig;
 
@@ -139,7 +140,8 @@ public class ElasticRepository implements ConciergeRepository {
         final var tenantQuery = TermQuery.of(p -> p.field(TENANT).value(getNormalizedValue(configKey.getTenantId())))._toQuery();
         final var configQuery = TermQuery.of(p -> p.field(CONFIG_NAME).value(getNormalizedValue(configKey.getConfigName())))._toQuery();
         final var configStateQuery = TermQuery.of(p -> p.field(CONFIG_STATE).value(getNormalizedValue(ConfigState.CREATED.name())))._toQuery();
-        final var searchQuery = BoolQuery.of(q -> q.must(List.of(orgQuery, namespaceQuery, tenantQuery, configQuery, configStateQuery)))._toQuery();
+        final var configTypeQuery = TermQuery.of(p -> p.field(CONFIG_TYPE).value(getNormalizedValue(configKey.getConfigType())))._toQuery();
+        final var searchQuery = BoolQuery.of(q -> q.must(List.of(orgQuery, namespaceQuery, tenantQuery, configQuery, configTypeQuery, configStateQuery)))._toQuery();
         final var searchResponse = client.search(SearchRequest.of(
                         s -> s.query(searchQuery)
                                 .requestCache(true)
@@ -161,11 +163,13 @@ public class ElasticRepository implements ConciergeRepository {
                 TermsQuery.of(q -> q.field(TENANT).terms(t -> t.value(getNormalizedValues(searchRequest.getTenants()))))._toQuery();
         final var configQuery = searchRequest.getConfigNames().isEmpty() ? MatchAllQuery.of(q -> q)._toQuery() :
                 TermsQuery.of(q -> q.field(CONFIG_NAME).terms(t -> t.value(getNormalizedValues(searchRequest.getConfigNames()))))._toQuery();
+        final var configTypeQuery = searchRequest.getConfigNames().isEmpty() ? MatchAllQuery.of(q -> q)._toQuery() :
+                TermsQuery.of(q -> q.field(CONFIG_TYPE).terms(t -> t.value(getNormalizedValues(searchRequest.getConfigTypes()))))._toQuery();
         final var configStateQuery = searchRequest.getConfigStates().isEmpty() ? MatchAllQuery.of(q -> q)._toQuery() :
                 TermsQuery.of(q -> q.field(CONFIG_STATE).terms(t -> t.value(getNormalizedValues(searchRequest.getConfigStates().stream().map(Enum::name).collect(Collectors.toSet()))))).
                         _toQuery();
         final var searchQuery = BoolQuery.of(q -> q.must(List.of(orgQuery, namespaceQuery, tenantQuery,
-                configQuery, configStateQuery)))._toQuery();
+                configQuery, configTypeQuery, configStateQuery)))._toQuery();
         return queryDetails(searchQuery, storedElasticRecordHit -> true);
     }
 
@@ -216,11 +220,12 @@ public class ElasticRepository implements ConciergeRepository {
     public void rollOverAndUpdate(ConfigDetails configDetails) {
         final var namespaceQuery = TermQuery.of(p -> p.field(NAMESPACE).value(getNormalizedValue(configDetails.getConfigKey().getNamespace())))._toQuery();
         final var configQuery = TermQuery.of(p -> p.field(CONFIG_NAME).value(getNormalizedValue(configDetails.getConfigKey().getConfigName())))._toQuery();
+        final var configTypeQuery = TermQuery.of(p -> p.field(CONFIG_TYPE).value(getNormalizedValue(configDetails.getConfigKey().getConfigType())))._toQuery();
         final var stateQuery = TermQuery.of(p -> p.field(CONFIG_STATE).value(getNormalizedValue(ConfigState.ACTIVATED.name())))._toQuery();
         final var orgQuery = TermQuery.of(p -> p.field(ORG).value(getNormalizedValue(configDetails.getConfigKey().getOrgId())))._toQuery();
         final var tenantQuery = TermQuery.of(p -> p.field(TENANT).value(getNormalizedValue(configDetails.getConfigKey().getTenantId())))._toQuery();
         final var searchQuery = BoolQuery.of(q -> q.must(List.of(orgQuery, namespaceQuery, tenantQuery,
-                configQuery, stateQuery)))._toQuery();
+                configQuery, configTypeQuery, stateQuery)))._toQuery();
         final var searchResponse = client.search(SearchRequest.of(
                         s -> s.query(searchQuery)
                                 .requestCache(true)